useThemeManager composable
